refactor(ramps): move resolveBuyAssetId into Ramp utils
Keep Buy CAIP resolution ownership with ramp intent helpers; TokenDetails only calls it from useHandleOnBuy.
