+++ app/components/UI/TokenDetails/hooks/useTokenAtomicActions.ts
@@ -1,12 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useNavigation } from '@react-navigation/native';
 import { useStore } from 'react-redux';
-import {
-  Hex,
-  CaipAssetType,
-  CaipChainId,
-  isCaipAssetType,
-} from '@metamask/utils';
+import { Hex, CaipAssetType, CaipChainId } from '@metamask/utils';
 import { strings } from '../../../../../locales/i18n';
 import Engine from '../../../../core/Engine';
 import { selectEvmChainId } from '../../../../selectors/networkController';
@@ -35,7 +30,6 @@ import { useSendNonEvmAsset } from '../../../hooks/useSendNonEvmAsset';
 import { formatChainIdToCaip } from '@metamask/bridge-controller';
 import { InitSendLocation } from '../../../Views/confirmations/constants/send';
 import { useSendNavigation } from '../../../Views/confirmations/hooks/useSendNavigation';
-import parseRampIntent from '../../Ramp/utils/parseRampIntent';
 import {
   getDetectedGeolocation,
   getOrders,
@@ -46,11 +40,11 @@ import {
   completedOrdersFromFiatOrders,
   completedOrdersFromRampsOrders,
 } from '../../Ramp/utils/determinePreferredProvider';
+import resolveBuyAssetId from '../../Ramp/utils/resolveBuyAssetId';
 import { BridgeToken } from '../../Bridge/types';
 import { adaptTokenSecurityData } from '../../Bridge/utils/tokenSecurityUtils';
 import { getSwapDestToken } from '../../Bridge/utils/getSwapDestToken';
 import { selectAssetsBySelectedAccountGroup } from '../../../../selectors/assets/assets-list';
-import { getCaipAssetIdForToken } from '../../Tokens/util/getCaipAssetIdForToken';
 import {
   isExploreTokenDetailsSource,
   TokenDetailsSource,
@@ -64,37 +58,6 @@ export type TokenActionInput = TokenI & {
   transactionActiveAbTests?: TransactionActiveAbTestEntry[];
   source?: TokenDetailsSource;
 };
-
-/**
- * Resolves the CAIP-19 asset id for Buy / ramp navigation.
- * Prefers an explicit caipAssetId, then Tokens-domain native/ERC-20 resolution
- * (`getCaipAssetIdForToken`), then legacy address+chain parseRampIntent.
- */
-export async function resolveBuyAssetId(
-  token: TokenActionInput,
-): Promise<string | undefined> {
-  try {
-    if (token.caipAssetId && isCaipAssetType(token.caipAssetId)) {
-      return token.caipAssetId;
-    }
-
-    const fromTokenModel = await getCaipAssetIdForToken(token);
-    if (fromTokenModel) {
-      return fromTokenModel;
-    }
-
-    if (isCaipAssetType(token.address)) {
-      return token.address;
-    }
-
-    return parseRampIntent({
-      chainId: getDecimalChainId(token.chainId as Hex),
-      address: token.address,
-    })?.assetId;
-  } catch {
-    return undefined;
-  }
-}
 
 interface BuySourceAsset {
   chainId: string;
